efi: earlycon: Add support for generic framebuffers and move to console subsystem
Add early console support for generic linear framebuffer devices.
This driver supports probing from cmdline early parameters
or from the device-tree using information in simple-framebuffer node.
The EFI functionality should be retained in whole.
The driver was disabled on ARM because of a bug in early_ioremap
implementation, and on IA64 and SH4 because of lack of
early_memremap_prot/early_ioremap on those architectures.

Signed-off-by: Markuss Broks <markuss.broks@gmail.com>